move fixture and method into conftest for shared usage
